+++ src/main/java/com/library/libraryapi/DAO/HireRepository.java
@@ -4,4 +4,5 @@ import com.library.libraryapi.Model.Hire;
 import org.springframework.data.repository.CrudRepository;
 
 public interface HireRepository extends CrudRepository<Hire, Long> {
+
 }